# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Multi-Merchant Social Commerce Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /merchants/{username}:
    get:
      operationId: getMerchant
      tags: [merchants]
      summary: Get merchant by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchants/{username}/products:
    get:
      operationId: getMerchantProducts
      tags: [merchants]
      summary: Get products for a specific merchant
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /merchants/{username}/highlights:
    get:
      operationId: getMerchantHighlights
      tags: [merchants]
      summary: Get highlights/collections for a specific merchant
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Highlights list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
  /merchant/me:
    get:
      operationId: getMyMerchantProfile
      tags: [merchants]
      summary: Get my merchant profile (authenticated)
      responses:
        "200":
          description: My merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No merchant profile yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/register:
    post:
      operationId: registerMerchant
      tags: [merchants]
      summary: Register as a merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMerchantBody"
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/products:
    get:
      operationId: listMyProducts
      tags: [merchants]
      summary: List my products (authenticated merchant)
      responses:
        "200":
          description: My product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProduct
      tags: [merchants]
      summary: Create a product (authenticated merchant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/products/{id}:
    put:
      operationId: updateProduct
      tags: [merchants]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [merchants]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/orders:
    get:
      operationId: getMerchantOrders
      tags: [merchants]
      summary: Get orders for my merchant (authenticated)
      responses:
        "200":
          description: Order list for merchant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /merchant/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [merchants]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Get current user's order history
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Checkout — create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Cart is empty or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SocialLink:
      type: object
      properties:
        platform:
          type: string
          description: "Platform name: instagram, facebook, tiktok, twitter, youtube"
        url:
          type: string
        handle:
          type: string
        followerCount:
          type: integer
      required:
        - platform
        - url
        - handle
    Merchant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        tagline:
          type: string
        avatar:
          type: string
        coverImage:
          type: string
        totalFollowers:
          type: integer
        totalSales:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isVerified:
          type: boolean
        website:
          type: string
        category:
          type: string
        location:
          type: string
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
        memberSince:
          type: string
      required:
        - id
        - userId
        - username
        - displayName
        - bio
        - avatar
        - totalFollowers
        - totalSales
        - isVerified
        - socialLinks
    RegisterMerchantBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        tagline:
          type: string
        bio:
          type: string
        category:
          type: string
        location:
          type: string
        website:
          type: string
      required:
        - username
        - displayName
        - bio
    Product:
      type: object
      properties:
        id:
          type: integer
        merchantId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        currency:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        isFeatured:
          type: boolean
        isSoldOut:
          type: boolean
        badge:
          type: string
        platform:
          type: string
          description: "Source platform: instagram, facebook, tiktok"
        postUrl:
          type: string
          description: "URL of the original social media post to embed"
        postedAt:
          type: string
        sellerUsername:
          type: string
        sellerAvatar:
          type: string
      required:
        - id
        - merchantId
        - title
        - description
        - price
        - currency
        - images
        - category
        - tags
        - likes
        - comments
        - shares
        - isFeatured
        - isSoldOut
        - platform
        - postedAt
        - sellerUsername
        - sellerAvatar
...
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        currency:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        badge:
          type: string
        platform:
          type: string
        postUrl:
          type: string
          description: "URL of the original social media post to embed"
      required:
        - title
        - description
        - price
        - images
        - category
        - tags
        - platform
...
    Highlight:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        coverImage:
          type: string
        category:
          type: string
        productCount:
          type: integer
      required:
        - id
        - title
        - coverImage
        - category
        - productCount
...
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerNotes:
          type: string
        address:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
      required:
        - customerName
        - address
    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        customerName:
          type: string
        customerNotes:
          type: string
        address:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        total:
          type: number
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - status
        - customerName
        - address
        - total
        - createdAt
        - items
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        merchantId:
          type: integer
        merchantUsername:
          type: string
        quantity:
          type: integer
        price:
          type: number
        title:
          type: string
        image:
          type: string
      required:
        - id
        - productId
        - merchantId
        - merchantUsername
        - quantity
        - price
        - title
    MerchantOrder:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        customerName:
          type: string
        customerNotes:
          type: string
        address:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        total:
          type: number
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - status
        - customerName
        - address
        - total
        - createdAt
        - items
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
      required:
        - status