# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Merchant Social Commerce Store API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateOrderBody {
  customerName: string;
  customerNotes?: string;
  address: string;
  locationLat?: number;
  locationLng?: number;
}