# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Merchant Social Commerce Store API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateProductBody {
  title: string;
  description: string;
  price: number;
  originalPrice?: number;
  currency?: string;
  images: string[];
  category: string;
  tags: string[];
  isFeatured?: boolean;
  badge?: string;
  platform: string;
  /** URL of the original social media post to embed */
  postUrl?: string;
}