# lib/api-zod/src/generated/types/highlight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Merchant Social Commerce Store API
 * OpenAPI spec version: 0.2.0
 */

export interface Highlight {
  id: number;
  title: string;
  coverImage: string;
  category: string;
  productCount: number;
}