# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Merchant Social Commerce Store API
 * OpenAPI spec version: 0.2.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  status: string;
  customerName: string;
  customerNotes?: string;
  address: string;
  locationLat?: number;
  locationLng?: number;
  total: number;
  createdAt: string;
  items: OrderItem[];
}